player animations, parameterize bullet render, changes to damage system, add flash to player when taking damage, synchronize life slider

+++ Proyecto Intermedio/Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void TakeDamage(float amount);
+    public void TakeDamage(int amount);
 }
